Fix C4244 warnings: change creationTime/lastUpdateTime from DWORD to unsigned long long
GetTickCount64() returns ULONGLONG, so member variables must match to avoid truncation warnings.

+++ UltraFastFileSearch-code/UltraFastFileSearch.cpp
@@ -2552,8 +2552,8 @@ class CProgressDialog : private CModifiedDialogImpl < CProgressDialog>, private 
 	WTL::CProgressBarCtrl progressBar;
 	bool canceled;
 	bool invalidated;
-	DWORD creationTime;
-	DWORD lastUpdateTime;
+	unsigned long long creationTime;
+	unsigned long long lastUpdateTime;
 	HWND parent;
 	std::tstring lastProgressText, lastProgressTitle;
 	bool windowCreated;
